+++ dotnet/Services/ICybersourceApi.cs
@@ -20,5 +20,6 @@ namespace Cybersource.Services
         Task<ConversionReportResponse> ConversionDetailReport(DateTime dtStartTime, DateTime dtEndTime);
         Task<string> RetrieveAvailableReports(DateTime dtStartTime, DateTime dtEndTime);
         Task<string> GetPurchaseAndRefundDetails(DateTime dtStartTime, DateTime dtEndTime);
+        Task<RetrieveTransaction> RetrieveTransaction(string transactionId);
     }
 }